feat: create ComponentShowcase page with UI component examples and interactions
